+++ gaffer/node/plugins.py
@@ -9,11 +9,11 @@ import sys
 from tornado import web
 
 class Plugin(object):
-    """ basic plugin interfacce """
+    """ basic plugin interface """
 
     name = ""
     version = "?"
-    descripton = ""
+    description = ""
     mandatory = []
 
     def app(self, cfg):