+++ gaffer/pm/commands/ps.py
@@ -11,10 +11,10 @@ from ...console_output import colored, GAFFER_COLORS
 
 class Ps(Command):
     """\
-        List your process informations
+        List your process information
         ------------------------------
 
-        Ps allows you to retrieve some process informations
+        Ps allows you to retrieve some process information
 
 
         .. image:: ../_static/gaffer_ps.png
@@ -29,7 +29,7 @@ class Ps(Command):
         Args
         ++++
 
-        *group*  is the name of the group of process recoreded in gafferd.
+        *group*  is the name of the group of process recorded in gafferd.
         By default it will be the name of your project folder.You can use
         ``.`` to specify the current folder.