+++ gaffer/process.py
@@ -276,7 +276,7 @@ class ProcessWatcher(object):
     def _start(self):
         if not self.active:
             self._timer = pyuv.Timer(self.loop)
-            # start the timer to refresh the informations
+            # start the timer to refresh the information
             # 0.1 is the minimum interval to fetch cpu stats for this
             # process.
             self._timer.start(self._async_refresh, 0.1, 0.1)
@@ -291,7 +291,7 @@ class Process(object):
     - **name**: name of the process
     - **cmd**: program command, string)
     - **args**: the arguments for the command to run. Can be a list or
-      a string. If **args** is  a string, it's splitted using
+      a string. If **args** is  a string, it's split using
       :func:`shlex.split`. Defaults to None.
     - **env**: a mapping containing the environment variables the command
       will run with. Optional
@@ -302,8 +302,8 @@ class Process(object):
       won't exit when the manager is stopped.
     - **shell**: boolean, run the script in a shell. (UNIX
       only)
-    - **redirect_output**: list of io to redict (max 2) this is a list of custom
-      labels to use for the redirection. Ex: ["a", "b"] will
+    - **redirect_output**: list of io to redirect (max 2) this is a list of
+      custom labels to use for the redirection. Ex: ["a", "b"] will
       redirect stdoutt & stderr and stdout events will be labeled "a"
     - **redirect_input**: Boolean (False is the default). Set it if
       you want to be able to write to stdin.
@@ -343,7 +343,7 @@ class Process(object):
                 self.cmd = args_[0]
                 self.args = args_[1:]
 
-        # replace envirnonnement variable in args
+        # replace environment variable in args
         # $PORT for example will become the given env variable.
         self.args = [substitute_env(arg, self.env) for arg in self.args]
 
@@ -459,7 +459,7 @@ class Process(object):
     @property
     def info(self):
         """ return the process info. If the process is monitored it
-        return the last informations stored asynchronously by the watcher"""
+        return the last information stored asynchronously by the watcher"""
 
         if not self._pprocess:
             self._pprocess = psutil.Process(self.pid)