+++ docs/news.rst
@@ -25,7 +25,7 @@ CHANGES
 2012/10/29 - version 0.4.0
 --------------------------
 
-- add environent variables support in the gafferd setting file.
+- add environment variables support in the gafferd setting file.
 - add a plugin system to easily extend :doc:`gafferd` using HTML sites
   or gaffer applications in Python